fix(db): #3 moved enums to ts files as they aren't supported in d.ts files
