+++ apps/framr-web/lib/types/enums/frame.enum.ts
@@ -1,0 +1,8 @@
+export enum FrameEnum {
+  UTIL = 'Utility',
+  RPT = 'Repeating',
+  SLD = 'Sliding',
+  ROT = 'Rotatory',
+  MTF = 'Magnetic Tool Phase',
+  GTF = 'Gravitational Tool Phase',
+}
