+++ apps/framr-web/lib/types/enums/tool.enum.ts
@@ -1,0 +1,4 @@
+export enum ToolEnum {
+  MWD = 'MWD',
+  LWD = 'LWD',
+}
