+++ apps/framr-web/lib/types/frame.d.ts
@@ -1,13 +1,5 @@
 import { DPoint } from './dpoint';
-
-export enum FrameEnum {
-  UTIL = 'Utility',
-  RPT = 'Repeating',
-  SLD = 'Sliding',
-  ROT = 'Rotatory',
-  MTF = 'Magnetic Tool Phase',
-  GTF = 'Gravitational Tool Phase',
-}
+import { FrameEnum } from './enums';
 
 export interface FramesetDpoint extends DPoint {
   isBaseInstance: boolean;
